requirements.txt
flask
transformers
torch
torchvision
torchaudio
pillow
